+++ mybatis-plus-core/src/test/java/com/baomidou/mybatisplus/core/toolkit/LambdaUtilsTest.java
@@ -1,7 +1,7 @@
 package com.baomidou.mybatisplus.core.toolkit;
 
-import com.baomidou.mybatisplus.core.toolkit.support.Property;
-import com.baomidou.mybatisplus.core.toolkit.support.SerializedLambda;
+import com.baomidou.mybatisplus.core.toolkit.support.lambda.SFunction;
+import com.baomidou.mybatisplus.core.toolkit.support.lambda.SerializedLambda;
 import lombok.Getter;
 import org.junit.Assert;
 import org.junit.Test;
@@ -42,7 +42,7 @@ public class LambdaUtilsTest {
 
         // 这个 TYPE 类型和 T 就没有关系了 ；
         // 如果有需要的话，使用 extends 来建立关系，保证class一致，稍微做点编写检查
-        <TYPE extends T> Cond<T> eq(Property<TYPE, ?> prop, Object val) {
+        <TYPE extends T> Cond<T> eq(SFunction<TYPE, ?> prop, Object val) {
             SerializedLambda lambda = LambdaUtils.resolve(prop);
             this.sb.append(lambda.getImplMethodName()).append(" = ").append(val);
             return this;
